refactor: refactor all components and page, create functions, utils, types
