+++ src/app/(private-pages)/agenda/eventListItem.tsx
@@ -1,0 +1,7 @@
+/* eslint-disable prettier/prettier */
+export default function EventListItem(){
+  return( <li className="flex bg-[#f8f9fa] px-4 py-2 gap-4 items-center border-l-[15px] border-red-500 rounded-lg mb-2">
+        <span className="text-md">08:00 - 10:00</span>
+        <h3 className="text-lg font-semibold">Título do Evento</h3>
+    </li>)
+}
