+++ src/app/(private-pages)/calendario/page.tsx
@@ -1,11 +1,11 @@
-import CalendarHeader from '../../../components/MonthCalendar/CalendarHeader';
-import MonthCalendar from '../../../components/MonthCalendar/MonthCalendar';
+import CalendarHeader from './MonthCalendar/CalendarHeader';
+import MonthCalendar from './MonthCalendar/MonthCalendar';
 
 export default function Calendar() {
   return (
-    <>
+    <div>
       <CalendarHeader />
       <MonthCalendar />
-    </>
+    </div>
   );
 }
