+++ src/app/(private-pages)/layout.tsx
@@ -1,4 +1,4 @@
-import Header from '../../components/header';
+import Header from '../../components/Header';
 import SideBarMenu from '../../components/SideBarMenu';
 
 export default function DashboardLayout({
@@ -10,7 +10,9 @@ export default function DashboardLayout({
     <body>
       <Header />
       <SideBarMenu />
-      <section className="mt-20 ml-52 py-8 px-16">{children}</section>
+      <section className="mt-20 ml-52 px-4 py-2 min-h-[calc(100vh-5rem)] bg-[#CED4DA]">
+        {children}
+      </section>
     </body>
   );
 }
