+++ src/app/globals.css
@@ -2,12 +2,30 @@
 @tailwind components;
 @tailwind utilities;
 
+::-webkit-scrollbar {
+  width: 5px;
+}
+
+::-webkit-scrollbar-track {
+  background: #b7e4c7;
+}
+
+::-webkit-scrollbar-thumb {
+    background: #2d6a4f;
+
+}
+
+::-webkit-scrollbar-thumb:hover {
+  background: #40916c;
+
+}
+
 @layer utilities {
-    .no-scrollbar::-webkit-scrollbar {
-        display: none;
-    }
-    .no-scrollbar {
-        -ms-overflow-style: none;  
-        scrollbar-width: none;  
+  .no-scrollbar::-webkit-scrollbar {
+    display: none;
+  }
+  .no-scrollbar {
+    -ms-overflow-style: none;
+    scrollbar-width: none;
   }
 }
