+++ src/app/layout.tsx
@@ -20,7 +20,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="pt-BR" className="light">
-      <body className={inter.className}>
+      <body className={inter.className} suppressHydrationWarning={true}>
         <CalendarContextWrapper>
           <Providers>{children}</Providers>
         </CalendarContextWrapper>
