+++ src/components/SideBarMenu.tsx
@@ -1,50 +1,23 @@
-// Use o Link nos componentes Button
-import { Button, ButtonGroup, Link } from '@nextui-org/react';
-import { BsListCheck } from 'react-icons/bs';
-import { FaRegCalendarCheck } from 'react-icons/fa';
-import { MdOutlineMenuBook } from 'react-icons/md';
-import { PiCalendarLight } from 'react-icons/pi';
+import { Button, ButtonGroup } from '@nextui-org/react';
+import NextLink from 'next/link';
+
+import { pagesRoutes } from '@/utils/pagesRoutes';
 
 export default function SideBarMenu() {
   return (
-    <aside className=" w-52 h-screen bg-[#31572C] fixed left-0 top-20">
-      {/*Como orientar verticalmente */}
-      <ButtonGroup className=" grid grid-cols-1 pt-5" radius="none" size="md">
-        <Button
-          startContent={<MdOutlineMenuBook />}
-          className="text-white bg-[#31572C] justify-start"
-          as={Link}
-          href="/cursos"
-        >
-          Meus cursos
-        </Button>
-
-        <Button
-          startContent={<PiCalendarLight />}
-          className="text-white bg-[#31572C] justify-start"
-          as={Link}
-          href="/calendario"
-        >
-          Calendario
-        </Button>
-
-        <Button
-          startContent={<BsListCheck />}
-          className="text-white bg-[#31572C] justify-start"
-          as={Link}
-          href="/agenda"
-        >
-          Agenda
-        </Button>
-
-        <Button
-          startContent={<FaRegCalendarCheck />}
-          className="text-white bg-[#31572C] justify-start"
-          as={Link}
-          href="/eventos"
-        >
-          Eventos
-        </Button>
+    <aside className=" w-52 h-screen bg-[#1B4332] fixed left-0 top-20">
+      <ButtonGroup className="grid grid-cols-1 pt-5" radius="none" size="md">
+        {pagesRoutes.map((route) => (
+          <Button
+            startContent={route.icon}
+            className="text-white bg-[#1B4332] justify-start"
+            as={NextLink}
+            href={route.path}
+            key={route.path}
+          >
+            {route.name}
+          </Button>
+        ))}
       </ButtonGroup>
     </aside>
   );
