--- src/components/eventListItem.tsx
@@ -1,7 +1,0 @@
-/* eslint-disable prettier/prettier */
-export default function EventListItem(){
-  return( <li className="flex bg-[#d9d9d9] px-4 py-2 gap-4 items-center border-l-[15px] border-red-500 rounded-lg mb-2">
-        <span className="text-md">08:00 - 10:00</span>
-        <h3 className="text-lg font-semibold">Título do Evento</h3>
-    </li>)
-}
