--- src/components/header.tsx
@@ -1,16 +1,0 @@
-import Perfil from './perfil';
-
-export default function Header() {
-  return (
-    <header className="flex bg-[#264422] justify-between items-center py-2 px-10 pr-3 fixed top-0 w-full h-20 z-20">
-      {/* logo da softex e input - direita, buscar evento */}
-      <a href="/">
-        <img
-          src="https://fap.softexrecife.org.br/assets/logo_softex_branca-8d8eb306ecdd1540cc4df965b41a811b428ef7fc9beec550d16902884ff9d117.png"
-          alt="logo"
-        />
-      </a>
-      <Perfil />
-    </header>
-  );
-}
