+++ src/types/events.ts
@@ -1,0 +1,9 @@
+import { Dayjs } from 'dayjs';
+
+export type Events = {
+  title: string;
+  location: string;
+  description: string;
+  day: Dayjs;
+  label: string;
+};
