+++ src/utils/courses.ts
@@ -1,0 +1,10 @@
+export const courses = [
+  {
+    title: 'Desenvolvedor Back end',
+    img: 'https://img.freepik.com/vetores-gratis/ilustracao-do-conceito-de-back-end_114360-22210.jpg?size=626&ext=jpg&ga=GA1.1.884335197.1703443415&semt=ais',
+  },
+  {
+    title: 'Desenvolvedor Front end',
+    img: 'https://img.freepik.com/vetores-gratis/ilustracao-do-conceito-de-programacao_114360-1351.jpg?size=626&ext=jpg&ga=GA1.1.884335197.1703443415&semt=ais',
+  },
+];
