+++ src/utils/eventsList.ts
@@ -1,0 +1,27 @@
+export const eventsList = [
+  {
+    title: 'Retorno as atividades presencias',
+    img: 'https://media.discordapp.net/attachments/1133474946560041173/1187470391162175659/WhatsApp_Image_2023-12-21_at_15.48.27.jpeg?ex=65a03b7b&is=658dc67b&hm=5983cd84b27bf6064cca5825691e0e57420f48dd209ec71f52e29c7216309ca9&=&format=webp&width=447&height=632',
+    date: '08/01/2023',
+  },
+  {
+    title: 'Webinar GitHub',
+    img: 'https://cdn.discordapp.com/attachments/1003202953647231007/1181992783536463924/git.jpeg?ex=65a7fd8f&is=6595888f&hm=45c4d610160a819fa7c4f3f359a01b30eb52e3c3bd2d1448de48ec357fceb368&',
+    date: '06/12/2023',
+  },
+  {
+    title: 'Webinar Metodologias Ágeis',
+    img: 'https://media.discordapp.net/attachments/1003202953647231007/1183833395986444459/banner.jpeg?ex=65a57544&is=65930044&hm=62808b55176ce4760ba74cafdabf5a9ababeac4dee05119be47cff0d37a00542&=&format=webp&width=1125&height=632',
+    date: '06/02/2023',
+  },
+  {
+    title: 'Inicio do Curso de FrontEnd',
+    img: 'https://i.pinimg.com/564x/bc/81/cf/bc81cfb9df56f9011f437ac32b0e71c8.jpg',
+    date: '06/08/2023',
+  },
+  {
+    title: 'Bem Vindos a Softex',
+    img: 'https://i.pinimg.com/564x/68/9c/a3/689ca3b9703d9d9a782267687fae3899.jpg',
+    date: '06/07/2023',
+  },
+];
