+++ src/utils/pagesRoutes.tsx
@@ -1,0 +1,27 @@
+import { BsListCheck } from 'react-icons/bs';
+import { FaRegCalendarCheck } from 'react-icons/fa';
+import { MdOutlineMenuBook } from 'react-icons/md';
+import { PiCalendarLight } from 'react-icons/pi';
+
+export const pagesRoutes = [
+  {
+    name: 'Eventos',
+    path: '/eventos',
+    icon: <FaRegCalendarCheck />,
+  },
+  {
+    name: 'Calendario',
+    path: '/calendario',
+    icon: <PiCalendarLight />,
+  },
+  {
+    name: 'Agenda',
+    path: '/agenda',
+    icon: <BsListCheck />,
+  },
+  {
+    name: 'Meus cursos',
+    path: '/cursos',
+    icon: <MdOutlineMenuBook />,
+  },
+];
